Fix GH#23652: Possible buffer overflow in `Capella::readString()`
'Backport` of #23654 (actually this one here existed first)

Update capella.cpp

+++ importexport/capella/capella.h
@@ -630,7 +630,7 @@ struct CapBracket {
 
 class Capella {
       static const char* errmsg[];
-      int curPos;
+      qint64 curPos;
 
       QFile* f;
       char* author;
